refactor(inbox): extract usePendingInvitations into hooks/
